+++ frontend/wailsjs/go/main/App.d.ts
@@ -3,6 +3,8 @@
 
 export function AppReady():Promise<void>;
 
+export function AutoDetectEnhanceEngine():Promise<Record<string, any>>;
+
 export function CloseWindow():Promise<void>;
 
 export function ExportPresets(arg1:string,arg2:string):Promise<string>;
@@ -12,6 +14,8 @@ export function FolderName(arg1:string):Promise<string>;
 export function GetAppVersion():Promise<string>;
 
 export function GetClipboardText():Promise<string>;
+
+export function GetGPUInfo():Promise<Record<string, any>>;
 
 export function ImportPresets():Promise<string>;